+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Upcoming
-* Food Heal Multiplier can now be configured even if Combat Test preset it enabled
+* `Food Heal Multiplier` can now be configured even if Combat Test preset it enabled
+* `Crops Require Water` no longer affects crops that aren't on farmland (e.g. Farmer's Delight rice)
 
 ## 2.11.7
 * Data packs